pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yolo_sam_training"
version = "0.1.0"
description = "Package for fine-tuning YOLO and SAM models using Label Studio outputs"
requires-python = ">=3.8"
dependencies = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
] 